Squash functionality ReadOnlyLedger hierarchy implementation into BaseLedger
changelog_begin
changelog_end
